2018-12-02 Implementando DAO  Implementei as classes de DAO, como foi pedido.  Também foram feitos mais ajustes nas libraries do projeto. Percebi que a classe do driver do PostgreSQL não era necessária estar no projeto, já que ela não era usada e nem se espera usar ela.

+++ ImplicareCore/src/br/cefetmg/implicare/model/serviceImpl/CargoManagementImpl.java
@@ -21,7 +21,7 @@ public class CargoManagementImpl implements CargoManagement {
     }
 
     @Override
-    public ArrayList<Cargo> listarCargoAreaEstudo(long cpf) throws PersistenceException, RemoteException {
+    public ArrayList<Cargo> listarCargoAreaEstudo(int cpf) throws PersistenceException, RemoteException {
         return dao.listarCargoAreaEstudo(cpf);
     }
